+++ rosbag2_transport/src/rosbag2_transport/record_options.cpp
@@ -40,7 +40,7 @@ Node convert<rosbag2_transport::RecordOptions>::encode(
     include_hidden_topics, include_unpublished_topics, ignore_leaf_topics,
     start_paused, use_sim_time, static_topics_uri, disable_keyboard_controls,
     statistics_max_publishing_rate, repeat_transient_local_messages,
-    repeat_all_transient_local_depth] = record_options;
+    repeat_all_transient_local_depth, topic_throttle_frequencies] = record_options;
   Node node;
   node["all_topics"] = all_topics;
   node["all_services"] = all_services;
@@ -78,6 +78,7 @@ Node convert<rosbag2_transport::RecordOptions>::encode(
   node["statistics_max_publishing_rate"] = statistics_max_publishing_rate;
   node["repeat_transient_local_messages"] = repeat_transient_local_messages;
   node["repeat_all_transient_local_depth"] = repeat_all_transient_local_depth;
+  node["topic_throttle_frequencies"] = topic_throttle_frequencies;
   return node;
 }
 
@@ -98,7 +99,7 @@ bool convert<rosbag2_transport::RecordOptions>::decode(
     include_hidden_topics, include_unpublished_topics, ignore_leaf_topics,
     start_paused, use_sim_time, static_topics_uri, disable_keyboard_controls,
     statistics_max_publishing_rate, repeat_transient_local_messages,
-    repeat_all_transient_local_depth] = record_options;
+    repeat_all_transient_local_depth, topic_throttle_frequencies] = record_options;
 
   optional_assign<bool>(node, "all_topics", all_topics);
   optional_assign<bool>(node, "all_services", all_services);
@@ -160,6 +161,8 @@ bool convert<rosbag2_transport::RecordOptions>::decode(
     node, "repeat_transient_local_messages", repeat_transient_local_messages);
   optional_assign<uint32_t>(node, "repeat_all_transient_local_depth",
     repeat_all_transient_local_depth);
+  optional_assign<std::unordered_map<std::string, double>>(
+    node, "topic_throttle_frequencies", topic_throttle_frequencies);
   return true;
 }